refactor: reorganize component structure and update import paths for consistency
