+++ src/pages/GamePage/TournamentPage/__mocks__/round2.ts
@@ -1,4 +1,4 @@
-import type { Match } from '@/api/types/game.ts';
+import type { Match } from '@/api/types/game';
 
 export const roundTwoData: Match = {
   id: 'final',
