+++ src/pages/GamePage/TournamentPage/_components/Chat/styles.css.ts
