+++ src/pages/GamePage/TournamentPage/_components/MatchLines/styles.css.ts
