+++ src/pages/GamePage/TournamentPage/_components/MatchNode/styles.css.ts
