+++ src/pages/GamePage/TournamentPage/_components/PlayerCard/index.tsx
