+++ src/pages/GamePage/TournamentPage/_components/PlayerCard/styles.css.ts
