+++ src/pages/GameSelectPage/GameMatchingPage/AutoMatchingPage/Auto1vs1/styles.css.ts
