+++ src/pages/GameSelectPage/GameMatchingPage/AutoMatchingPage/AutoTournament/styles.css.ts
