+++ src/pages/GameSelectPage/GameMatchingPage/_components/invite-friend-dialog/styles.css.ts
