+++ src/pages/GameSelectPage/GameMatchingPage/_components/user-card/styles.css.ts
