+++ src/pages/GameSelectPage/GameMatchingPage/_components/waiting-message/index.tsx
